adds 3d landmarks widget
