+++ menpowidgets/items.py
@@ -16,11 +16,6 @@ def view_widget(items, **kwargs):
     kwargs : optional
         Keyword arguments that will be forwarded to the appropriate widget.
 
-    Returns
-    -------
-    widget
-        The appropriate instantiated widget.
-
     Raises
     ------
     ValueError
@@ -42,9 +37,15 @@ def view_widget(items, **kwargs):
     if isinstance(template, PointCloud) and template.n_dims == 2:
         from .base import visualize_shapes_2d
         visualize_shapes_2d(items, **kwargs)
-    elif isinstance(template, LandmarkManager):
+    elif isinstance(template, PointCloud) and template.n_dims == 3:
+        from .base import visualize_shapes_3d
+        visualize_shapes_3d(items, **kwargs)
+    elif isinstance(template, LandmarkManager) and template.n_dims == 2:
         from .base import visualize_landmarks_2d
         visualize_landmarks_2d(items, **kwargs)
+    elif isinstance(template, LandmarkManager) and template.n_dims == 3:
+        from .base import visualize_landmarks_3d
+        visualize_landmarks_3d(items, **kwargs)
     elif isinstance(template, Image):
         from .base import visualize_images
         visualize_images(items, **kwargs)
